Phase4: extract pure builders into phase4_builders (no behavior change)
